+++ packages/@kagal-taistamp/src/__tests__/compat.mjs
@@ -12,8 +12,10 @@
 import {
   asLeapSeconds,
   asNonce,
+  asSignature,
   composeSignaturePayload,
   extractLeapSeconds,
+  extractSignature,
   newEd25519Signer,
   newNonce,
   newTaistampHandler,
@@ -131,7 +133,9 @@ checkFunction('composeSignaturePayload', composeSignaturePayload);
 checkFunction('asLeapSeconds', asLeapSeconds);
 checkFunction('asNonce', asNonce);
 checkFunction('newNonce', newNonce);
+checkFunction('asSignature', asSignature);
 checkFunction('extractLeapSeconds', extractLeapSeconds);
+checkFunction('extractSignature', extractSignature);
 checkFunction('tai64nLabelFromUTC (main)', tai64nLabelFromUTCMain);
 
 // /utils subpath