+++ packages/@kagal-taistamp/src/index.ts
@@ -42,4 +42,5 @@ export {
   newNonce,
   type Nonce,
 } from './nonce';
+export { asSignature, extractSignature } from './signature';
 export { tai64nLabelFromUTC } from './time';